feat(api): auth module finished

+++ api/test/groups/users-test.group.ts
@@ -2,22 +2,27 @@ import { INestApplication } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { AuthModule } from '../../src/auth/auth.module';
 import * as request from 'supertest';
 import { Repository } from 'typeorm';
-import { PostgresTypeOrmConfigFactory } from '../src/config/postgres-typeorm.config';
-import { CreateUserDto } from '../src/users/dto/create-user.dto';
-import { UpdateUserDto } from '../src/users/dto/update-user.dto';
-import { User } from '../src/users/entities/user.entity';
-import { UsersModule } from '../src/users/users.module';
-import { UsersMock } from './mocks/users.mock';
+import { PostgresTypeOrmConfigFactory } from '../../src/config/postgres-typeorm.config';
+import { CreateUserDto } from '../../src/users/dto/create-user.dto';
+import { UpdateUserDto } from '../../src/users/dto/update-user.dto';
+import { User } from '../../src/users/entities/user.entity';
+import { UsersModule } from '../../src/users/users.module';
+import { UsersMock } from '../mocks/users.mock';
+import { AuthService } from '../../src/auth/auth.service';
 
-describe('Users', () => {
+export default () => {
   let app: INestApplication;
-  let repository: Repository<User>;
+  let userRepository: Repository<User>;
+  let authService: AuthService;
+  let access_token: string;
 
   beforeAll(async () => {
     const module = await Test.createTestingModule({
       imports: [
+        AuthModule,
         UsersModule,
         ConfigModule.forRoot(),
         TypeOrmModule.forRootAsync({
@@ -31,7 +36,17 @@ describe('Users', () => {
     app = module.createNestApplication();
     await app.init();
 
-    repository = module.get('UserRepository');
+    userRepository = module.get('UserRepository');
+    authService = module.get<AuthService>(AuthService);
+
+    const admin = await userRepository.save({
+      name: 'Admin Test E2E',
+      email: 'admin_test@e2e.com',
+      password: 'admin_password',
+    });
+
+    const response = await authService.login(admin);
+    access_token = response.access_token;
   });
 
   describe('/POST Users', () => {
@@ -39,7 +54,8 @@ describe('Users', () => {
       const createUserDto: CreateUserDto = UsersMock.create();
 
       return request(app.getHttpServer())
-        .post('/users')
+        .post('/api/v1/users')
+        .set('Authorization', `Bearer ${access_token}`)
         .send(createUserDto)
         .expect(201);
     });
@@ -48,7 +64,8 @@ describe('Users', () => {
       const createUserDto: CreateUserDto = UsersMock.create();
 
       return request(app.getHttpServer())
-        .post('/users')
+        .post('/api/v1/users')
+        .set('Authorization', `Bearer ${access_token}`)
         .send(createUserDto)
         .expect(400)
         .expect({
@@ -64,7 +81,8 @@ describe('Users', () => {
       const updateUserDto: UpdateUserDto = UsersMock.update();
 
       return request(app.getHttpServer())
-        .put(`/users/1256`)
+        .put(`/api/v1/users/1256`)
+        .set('Authorization', `Bearer ${access_token}`)
         .send(updateUserDto)
         .expect(404)
         .expect({
@@ -75,7 +93,7 @@ describe('Users', () => {
     });
 
     it('deve atualizar as informações do usuário', async () => {
-      const user = <User>await repository.findOne({
+      const user = <User>await userRepository.findOne({
         where: {
           email: UsersMock.CREATE_EMAIL,
         },
@@ -84,7 +102,8 @@ describe('Users', () => {
       const updateUserDto: UpdateUserDto = UsersMock.update();
 
       return request(app.getHttpServer())
-        .put(`/users/${user.id}`)
+        .put(`/api/v1/users/${user.id}`)
+        .set('Authorization', `Bearer ${access_token}`)
         .send(updateUserDto)
         .expect(200);
     });
@@ -93,10 +112,11 @@ describe('Users', () => {
       const createUserDto: CreateUserDto = UsersMock.create();
       const updateUserDto: UpdateUserDto = UsersMock.update();
 
-      const user = await repository.save(createUserDto);
+      const user = await userRepository.save(createUserDto);
 
       return request(app.getHttpServer())
-        .put(`/users/${user.id}`)
+        .put(`/api/v1/users/${user.id}`)
+        .set('Authorization', `Bearer ${access_token}`)
         .send(updateUserDto)
         .expect(400)
         .expect({
@@ -110,20 +130,22 @@ describe('Users', () => {
   describe('/GET Users', () => {
     it('deve retornar uma lista de usuários', async () => {
       return request(app.getHttpServer())
-        .get('/users')
+        .get('/api/v1/users')
+        .set('Authorization', `Bearer ${access_token}`)
         .expect('Content-Type', /json/)
         .expect(200);
     });
 
     it('deve retornar um usuário pelo id', async () => {
-      const user = <User>await repository.findOne({
+      const user = <User>await userRepository.findOne({
         where: {
           email: UsersMock.CREATE_EMAIL,
         },
       });
 
       return request(app.getHttpServer())
-        .get(`/users/${user.id}`)
+        .get(`/api/v1/users/${user.id}`)
+        .set('Authorization', `Bearer ${access_token}`)
         .expect(200)
         .expect('Content-Type', /json/);
     });
@@ -131,22 +153,23 @@ describe('Users', () => {
 
   describe('/DELETE Users', () => {
     it('deve remover um usuário', async () => {
-      const user = <User>await repository.findOne({
+      const user = <User>await userRepository.findOne({
         where: {
           email: UsersMock.UPDATE_EMAIL,
         },
       });
 
       return request(app.getHttpServer())
-        .delete(`/users/${user.id}`)
+        .delete(`/api/v1/users/${user.id}`)
+        .set('Authorization', `Bearer ${access_token}`)
         .expect(200);
     });
   });
 
   afterAll(async () => {
-    await repository.query(
-      `DELETE FROM users WHERE email = '${UsersMock.CREATE_EMAIL}'`,
+    await userRepository.query(
+      `DELETE FROM users WHERE email in ('${UsersMock.CREATE_EMAIL}', 'admin_test@e2e.com')`,
     );
     await app.close();
   });
-});
+};
